tflite seems to work

+++ Src/Representations/Infrastructure/R2KAudioData.h
@@ -14,5 +14,5 @@
 
 STREAMABLE(R2KAudioData,
 {,
-    (int) dummyData,
+    (float) dummyData,
 });
